+++ src/main/webapp/components/user-settings/hidden-results-settings.js
@@ -11,9 +11,9 @@ import VisibilityIcon from '@material-ui/icons/Visibility'
 import gql from 'graphql-tag'
 import { Map } from 'immutable'
 import React from 'react'
-import Inspector from '../components/inspector'
-import { useApolloFallback } from '../react-hooks'
-import ErrorMessage from '../components/network-retry/inline-retry'
+import Inspector from '../inspector'
+import { useApolloFallback } from '../../react-hooks'
+import ErrorMessage from '../network-retry/inline-retry'
 
 const resultQuery = gql`
   query ResultQuery($ids: [ID]!) {
