+++ src/main/webapp/components/user-settings/source-select.js
@@ -9,8 +9,8 @@ import HomeIcon from '@material-ui/icons/Home'
 import WarningIcon from '@material-ui/icons/Warning'
 import { get } from 'immutable'
 import React from 'react'
-import OutlinedSelect from '../input/outlined-select'
-import { useApolloFallback } from '../react-hooks'
+import OutlinedSelect from '../../input/outlined-select'
+import { useApolloFallback } from '../../react-hooks'
 import gql from 'graphql-tag'
 
 const query = gql`
