+++ src/main/webapp/components/user-settings/theme-settings.js
