+++ src/main/webapp/components/user-settings/user-settings.stories.js
@@ -2,8 +2,8 @@ import { action } from '@connexta/ace/@storybook/addon-actions'
 import { boolean } from '@connexta/ace/@storybook/addon-knobs'
 import { Map } from 'immutable'
 import React from 'react'
-import { storiesOf } from '../@storybook/react'
-import useState from '../@storybook/use-state'
+import { storiesOf } from '../../@storybook/react'
+import useState from '../../@storybook/use-state'
 import HiddenResultsSettings from './hidden-results-settings'
 import NotificationSettings from './notification-settings'
 import SearchSettings from './search-settings'
